fix: изменение типа telegram_id с INT на BIGINT
- Изменен тип колонки telegram_id с Integer на BigInteger в моделях User и TelegramVerificationCode
- Создана миграция fix_telegram_id_bigint.py для изменения типа колонки в БД
- Обновлена миграция add_telegram_notifications.py для создания колонки сразу с BIGINT
- Обновлен apply_migrations.py для автоматического применения новой миграции
- Исправлена ошибка 'Out of range value for column telegram_id' для больших Telegram ID (> 2,147,483,647)

+++ auth/models/user_models.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String, Enum, DateTime, Boolean
+from sqlalchemy import Column, Integer, BigInteger, String, Enum, DateTime, Boolean
 from sqlalchemy.sql import func
 from sqlalchemy.orm import relationship
 from database import Base
@@ -18,7 +18,7 @@ class User(Base):
     email = Column(String(100), unique=True, index=True, nullable=False)
     hashed_password = Column(String(255), nullable=False)
     role = Column(Enum(UserRole), nullable=False, default=UserRole.CUSTOMER)  # Оставляем для обратной совместимости
-    telegram_id = Column(Integer, nullable=True, unique=True, index=True)  # Telegram ID пользователя
+    telegram_id = Column(BigInteger, nullable=True, unique=True, index=True)  # Telegram ID пользователя
     is_active = Column(Boolean, default=True)
     created_at = Column(DateTime(timezone=True), server_default=func.now())
     updated_at = Column(DateTime(timezone=True), onupdate=func.now())
